requirements.txt
aiogram==3.4.1
requests==2.31.0
cerebras-cloud-sdk==1.46.0
python-dotenv==1.0.0
openai==1.12.0
tavily-python==0.3.3
supabase==2.5.0
apscheduler==3.10.4
aiohttp==3.9.1
beautifulsoup4==4.12.2
PyPDF2==3.0.1
numpy==1.26.2
asyncio-throttle==1.0.2
python-dateutil==2.8.2
pytz==2023.3
